+++ options/pin.js
@@ -1,10 +1,10 @@
-import { getPackageInformation, savePackageManager } from '../helpers/files.js'
+import { getPackageJson, savePackageJson } from '../helpers/files.js'
 
 export const pinPackageManager = async (packageCommand) => {
-  const packageJson = getPackageInformation()
+  const packageJson = getPackageJson()
   if (packageJson) {
     packageJson.swpm = packageCommand
-    await savePackageManager(packageJson)
+    await savePackageJson(packageJson)
   }
   process.exit(1)
 }