+++ packages/managers/npm.js
@@ -1,5 +1,6 @@
 const npm = {
   cmd: 'npm',
+  color: '#e12b3b',
   url: 'https://www.npmjs.com/',
   version: '',
   lockFile: 'package-lock.json'