+++ packages/managers/pnpm.js
@@ -1,5 +1,6 @@
 const pnpm = {
   cmd: 'pnpm',
+  color: '#f7ad24',
   url: 'https://pnpm.io/',
   version: '',
   lockFile: 'pnpm-lock.yaml'