+++ packages/managers/yarn.js
@@ -1,5 +1,6 @@
 const yarn = {
   cmd: 'yarn',
+  color: '#3a90b8',
   url: 'https://yarnpkg.com/',
   version: '',
   lockFile: 'yarn.lock'